A property with idsm:referenceByUri true must not have a class check in the respective SHACL shape, for instance: ids:theme and DigitalContentShape
Reason:
the serialised output of referenceByUri is the plain URI. A SHACL validator however searches for a certain class assertion, which it will not find. Therefore, it will through a validation error. However, the object is an intance of the class. This information is just not known to the validator.