Removing minCount 1 for actionRefinement
As the Codes of Action do not have an actionRefinement defined, we cannot have an sh:minCount 1 in the shapes. Otherwise, the Information model is violating against its own shapes.
This is not meant as a "final solution", but as a hotfix. Adding actionRefinements to the codes also sounds like a legitimate approach.